Fix support for remote modules
I had not used this functionality in a while; it turns out Terraform refuses
to initialize a remote workspace when there are files in place. Unforunately it
also sometimes needs those files. I think we will need a better solution in
future - e.g. using go-getter or similar to separately fetch the module.

Signed-off-by: Nic Cope <negz@rk0n.org>